Move the models to their own files
Refs 140
