+++ parking_permits_app/models/__init__.py
@@ -1,0 +1,29 @@
+from .address import Address
+from .company import Company
+from .contract_type import ContractType
+from .customer import Customer
+from .driving_class import DrivingClass
+from .driving_licence import DrivingLicence
+from .low_emission_criteria import LowEmissionCriteria
+from .parking_permit import ParkingPermit
+from .parking_zone import ParkingZone
+from .product import Product
+from .product_price import ProductPrice
+from .vehicle import Vehicle
+from .vehicle_type import VehicleType
+
+__all__ = [
+    "Address",
+    "Company",
+    "ContractType",
+    "Customer",
+    "DrivingClass",
+    "DrivingLicence",
+    "LowEmissionCriteria",
+    "ParkingPermit",
+    "ParkingZone",
+    "Product",
+    "ProductPrice",
+    "Vehicle",
+    "VehicleType",
+]
