+++ parking_permits_app/models/mixins.py
@@ -1,0 +1,21 @@
+import uuid
+
+from django.contrib.gis.db import models
+from django.utils.translation import gettext_lazy as _
+
+
+class TimestampedModelMixin(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True, verbose_name=_("Time created"))
+    modified_at = models.DateTimeField(auto_now=True, verbose_name=_("Time modified"))
+
+    class Meta:
+        abstract = True
+
+
+class UUIDPrimaryKeyMixin(models.Model):
+    id = models.UUIDField(
+        primary_key=True, unique=True, editable=False, default=uuid.uuid4
+    )
+
+    class Meta:
+        abstract = True
