+++ parking_permits_app/models/parking_zone.py
@@ -1,0 +1,20 @@
+from django.conf import settings
+from django.contrib.gis.db import models
+from django.utils.translation import gettext_lazy as _
+
+from .mixins import TimestampedModelMixin, UUIDPrimaryKeyMixin
+
+
+class ParkingZone(TimestampedModelMixin, UUIDPrimaryKeyMixin):
+    name = models.CharField(_("Name"), max_length=128, blank=False, null=False)
+    location = models.MultiPolygonField(
+        _("Area (2D)"), srid=settings.SRID, blank=False, null=False
+    )
+
+    class Meta:
+        db_table = "parking_zone"
+        verbose_name = _("Parking zone")
+        verbose_name_plural = _("Parking zones")
+
+    def __str__(self):
+        return "%s - %s" % (self.id, self.name)
